WIP: Move all infrastructure code to `infra/` dir
