+++ test/screenshot/infra/auth/.gitignore
