+++ test/screenshot/infra/commands/approve.js
