+++ test/screenshot/infra/commands/build.js
