+++ test/screenshot/infra/commands/clean.js
