+++ test/screenshot/infra/commands/demo.js
