+++ test/screenshot/infra/commands/proto.js
