+++ test/screenshot/infra/commands/test.js
