+++ test/screenshot/infra/lib/cbt-api.js
