+++ test/screenshot/infra/lib/cli.js
