+++ test/screenshot/infra/lib/cloud-storage.js
