+++ test/screenshot/infra/lib/constants.js
