+++ test/screenshot/infra/lib/controller.js
