+++ test/screenshot/infra/lib/duration.js
