+++ test/screenshot/infra/lib/externs.js
