+++ test/screenshot/infra/lib/file-cache.js
