+++ test/screenshot/infra/lib/git-repo.js
