+++ test/screenshot/infra/lib/golden-file.js
