+++ test/screenshot/infra/lib/golden-io.js
