+++ test/screenshot/infra/lib/image-cropper.js
