+++ test/screenshot/infra/lib/local-storage.js
