+++ test/screenshot/infra/lib/logger.js
