+++ test/screenshot/infra/lib/report-writer.js
