+++ test/screenshot/infra/lib/selenium-api.js
