+++ test/screenshot/infra/proto/cbt.pb.js
