+++ test/screenshot/infra/proto/cbt.proto
