+++ test/screenshot/infra/proto/github.proto
