+++ test/screenshot/infra/proto/mdc.pb.js
