+++ test/screenshot/infra/proto/mdc.proto
