+++ test/screenshot/infra/proto/selenium.pb.js
