+++ test/screenshot/infra/proto/selenium.proto
