+++ test/screenshot/report/report.hbs
@@ -19,7 +19,7 @@
 
     <!-- TODO(acdvorak): Download clipboard.js and browser icons to the repo instead of using a CDN -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/clipboard.js/2.0.0/clipboard.min.js"></script>
-    <script src="../lib/duration.js"></script>
+    <script src="../infra/lib/duration.js"></script>
     <script src="../out/report/report.js"></script>
   </body>
 </html>
